+++ packages/manager/src/features/linodes/LinodesCreate/types.ts
@@ -7,7 +7,7 @@ import {
 import { Region } from '@linode/api-v4/lib/regions';
 import { StackScript, UserDefinedField } from '@linode/api-v4/lib/stackscripts';
 import { APIError } from '@linode/api-v4/lib/types';
-import { Tag } from 'src/components/TagsInput';
+import { Tag } from 'src/components/TagsInput/TagsInput';
 import { ExtendedType } from 'src/utilities/extendType';
 
 export interface ExtendedLinode extends Linode {